client/requirements.txt
google-cloud-storage>=2.16
google-cloud-pubsub>=2.21
Pillow>=10.0
python-dotenv>=1.0
